console: Change location of invitations logics and rename store/*/user
