+++ pkg/webui/console/containers/applications-table/index.js
@@ -44,7 +44,7 @@ import {
   getApplicationsList,
 } from '@console/store/actions/applications'
 
-import { selectUserIsAdmin } from '@console/store/selectors/user'
+import { selectUserIsAdmin } from '@console/store/selectors/logout'
 import {
   selectApplications,
   selectApplicationsTotalCount,
