+++ pkg/webui/console/containers/gateways-table/index.js
@@ -37,7 +37,7 @@ import { checkFromState, mayCreateGateways } from '@console/lib/feature-checks'
 
 import { getGatewaysList, restoreGateway, deleteGateway } from '@console/store/actions/gateways'
 
-import { selectUserIsAdmin } from '@console/store/selectors/user'
+import { selectUserIsAdmin } from '@console/store/selectors/logout'
 import {
   selectGateways,
   selectGatewaysTotalCount,
