+++ pkg/webui/console/store/actions/users.js
@@ -73,3 +73,29 @@ export const [
     failure: getUsersRightsListFailure,
   },
 ] = createGetRightsListRequestActions(SHARED_NAME)
+
+export const GET_USER_INVITATIONS_BASE = createPaginationBaseActionType('INVITATIONS')
+export const [
+  {
+    request: GET_USER_INVITATIONS,
+    success: GET_USER_INVITATIONS_SUCCESS,
+    failure: GET_USER_INVITATIONS_FAILURE,
+  },
+  {
+    request: getUserInvitations,
+    success: getUserInvitationsSuccess,
+    failure: getUserInvitationsFailure,
+  },
+] = createPaginationRequestActions('INVITATIONS')
+
+export const SEND_INVITE_BASE = 'SEND_INVITE'
+export const [
+  { request: SEND_INVITE, success: SEND_INVITE_SUCCESS, failure: SEND_INVITE_FAILURE },
+  { request: sendInvite, success: sendInviteSuccess, failure: sendInviteFailure },
+] = createRequestActions(SEND_INVITE_BASE, email => ({ email }))
+
+export const DELETE_INVITE_BASE = 'DELETE_INVITE'
+export const [
+  { request: DELETE_INVITE, success: DELETE_INVITE_SUCCESS, failure: DELETE_INVITE_FAILURE },
+  { request: deleteInvite, success: deleteInviteSuccess, failure: deleteInviteFailure },
+] = createRequestActions(DELETE_INVITE_BASE, email => ({ email }))
